postgres_to_es/pyproject.toml
[tool.poetry]
name = "postgres-to-es"
version = "0.1.0"
description = "Practicum sprint 3 task"
authors = ["iManapov <manapovilnar@gmail.com>"]
readme = "README.md"
packages = [{include = "postgres_to_es"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "2.9.3"
python-dotenv = "^0.21.0"
pydantic = "^1.10.2"
elasticsearch = "7.17.6"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
isort = "^5.10.1"
black = "^22.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
